requirements/dev.txt
asgiref==3.8.1
autopep8==2.3.1
dj-database-url==2.3.0
Django==5.1.3
django-autoslug==1.9.9
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-silk==5.3.1
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gprof2dot==2024.6.6
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.10.0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
